Move classes out to their own files.
