+++ ruby/lib/ugly_trivia/categorized_questions.rb
@@ -1,0 +1,15 @@
+module UglyTrivia
+  class CategorizedQuestions
+    def initialize(categories)
+      @questions = Hash[categories.map { |c| [c, []] }]
+    end
+
+    def add(category, question)
+      @questions[category] << question
+    end
+
+    def next_question(category)
+      @questions[category].shift
+    end
+  end
+end
