+++ ruby/lib/ugly_trivia/console_output.rb
@@ -1,0 +1,7 @@
+module UglyTrivia
+  class ConsoleOutput
+    def write(line)
+      puts line
+    end
+  end
+end
