+++ ruby/lib/ugly_trivia/penalty_box.rb
@@ -1,0 +1,15 @@
+module UglyTrivia
+  class PenaltyBox
+    def initialize
+      @players = Set.new
+    end
+
+    def hold(player)
+      @players << player
+    end
+
+    def holding?(player)
+      @players.include?(player)
+    end
+  end
+end
