+++ ruby/lib/ugly_trivia/purse.rb
@@ -1,0 +1,19 @@
+module UglyTrivia
+  class Purse
+    def initialize
+      @coins = 0
+    end
+
+    def add_coin
+      @coins += 1
+    end
+
+    def total
+      @coins
+    end
+
+    def to_s
+      @coins.to_s
+    end
+  end
+end
